requirements.txt
Django==4.1
djangorestframework
djangorestframework-simplejwt
requests
gunicorn
psycopg2-binary
dj-database-url